54.6 Only authorized users can delete comments. They must be admins or they must be the author of that comment

+++ app/Http/Livewire/DeleteComment.php
@@ -19,9 +19,9 @@ class DeleteComment extends Component {
 
     public function deleteComment(){
         #Authorization
-        //if(auth()->guest() || auth()->user()->cannot('update', $this->comment)){
-        //abort(Response::HTTP_FORBIDDEN);
-        //}
+        if(auth()->guest() || auth()->user()->cannot('delete', $this->comment)){
+            abort(Response::HTTP_FORBIDDEN);
+        }
 
         Comment::destroy($this->comment->id);
